+++ continuousflex/viewers/__init__.py
@@ -28,7 +28,6 @@ from .viewer_nma_dimred import FlexDimredNMAViewer
 from .viewer_structure_mapping import FlexProtStructureMappingViewer
 from .viewer_subtomograms_synthesize import FlexProtSynthesizeSubtomoViewer
 from .viewer_pdb_dimred import FlexProtPdbDimredViewer, VolumeTrajectoryViewer
-from .viewer_pca_pdbs import FlexProtPCAPdbViewer
 from .viewer_subtomograms_classify import FlexProtSubtomoClassifyViewer
 from .viewer_nma_alignment_vol import FlexAlignmentNMAVolViewer
 from .viewer_nma_dimred_vol import FlexDimredNMAVolViewer
